requirements.txt
flask
pywebview
PyPDF2
spire.pdf
pyinstaller
